Move parsed source/trigger types to separate files
This will make it easier to split source/trigger parsing into separate
files, and removes unnecessary dependencies on parsing.
