+++ ts/src/header-validator/aggregatable-contributions.ts
@@ -1,12 +1,11 @@
 import { SourceType } from '../source-type'
 import * as filters from './filters'
+import { AggregationKeys, FilterData } from './source'
 import {
   AggregatableTriggerDatum,
   AggregatableValues,
   AggregatableValuesConfiguration,
-  AggregationKeys,
-  FilterData,
-} from './validate-json'
+} from './trigger'
 
 export type AggregatableContribution = {
   key: bigint
